mathastext 1.15 2012-09-26
The mathastext package automatically (or only locally)
changes the document fonts used in math mode for letters and
digits (and a few other punctuation and symbol signs) to
replace them by the font used for the text. Thus, the
package makes it possible to use a quite arbitrary font
without worrying too much that it does not have specially
designed accompanying math fonts. Also, mathastext provides
a simple mechanism in order to use more than one
math-as-text font in the same document.

New with 1.15: with the `subdued' option the action of
mathastextification has be activated explicitely, locally in
the document.