mathastext 1.15c 2012-10-05
Optimal typographical results for documents containing
mathematical symbols can only be hoped for with math fonts
specifically designed to match a given text typeface.
Although the list of freely available math fonts (alongside
the Computer Modern and AMS extension fonts) is slowly
expanding (fourier, kpfonts, mathdesign, pxfonts, txfonts,
and others...) it remains limited, and the situation is even
worse with Unicode fonts (XeTeX/LuaTeX). So if you can't
find a math font which fits well with your favorite text
font, and wish to still be able to typeset mathematical
documents, perhaps not of the highest typographical quality,
but at least not subjected to obvious visual incompatibi-
-lities between your text font and the math fonts,
try out mathastext: it will simply use the text font also
for the math!

New with 1.15c:

  + it is now possible to use multiple distinct LGR encoded
  fonts for the Greek letters in math mode, one for each math
  version.

  + improved documentation.

New with 1.15:

  + with the subdued option the action of mathastext can be
  limited to a portion of your document.