mathastext 1.3 2013-09-02
1.3 [2013/09/02]

The main change is the command \MTsetmathskips which allows to
specify, on a letter by letter basis, extra skips to be
automatically inserted before and after each occurence of the
given letter in math mode. This allows to fine tune the use of
a given text font. A random example:

   \MTsetmathskips{x}{0.4mu}{0.2mu}

Also, the functionality of the package under lualatex is now
at the same level as under xelatex; requires the lualatex as
distributed with TL2013 or later.

The documentation is better hyperlinked.