mathatext 1.3a 2013-09-04
1.3a [2013/09/04]

The command \MTsetmathskips from recent release 1.3 is more
flexible (e.g. for use in an \@for loop): its first argument
may be a macro (or any expandable code) giving the letter.

The amsmath macro \resetMathstrut@ is replaced only when
strictly necessary (i.e. if the user makes the parentheses
mathematically active via \MTnonlettersobeymathxx).

Again some improvements in the documentation.