mathastext 1.3c 2013-12-14
1.3c [2013/12/14]

   * added a starred variant to \MTversion which tells mathastext to
only do the math set-up and not modify the text fonts.

   * added second optional version name argument to \Mathastext and to
\MTDeclareVersion, to transfer settings for things not otherwise changed
by mathastext from a math version to the one declared. This is mainly
for symbols and large symbols to be the bold ones when the user sets up
the series of a mathastextified font to be bold in a mathastext-declared
version.

1.3b [2013/12/12]

Archival note:

This 1.3b release had a bug in its modified \MTDeclareVersion.
1.3c fixed it and added the starred variant to \MTversion.
At that time the author was not yet using Git version control
for this project, and the 1.3b zip submission to CTAN was
not kept.