mathastext 1.3n 2016-04-22
At long last, mathastext takes care properly of annoying and
perplexing amsmath's \newmcodes@. The very recent change in
amsopn.sty finally made it compatible with Unicode engines, but
anyhow, mathastext must do its own patch to use the correct font.
All of this taking into account the various options passed to
the package. Lots of trouble for a tiny thing.

(a typo in the announcement was corrected at time of archival here)