mathastext 1.3o 2016-05-03
mathastext fixes an issue related to a feature of LuaLaTeX and
luaotfload that OpenType fonts are declared in one of two modes:
node and base, and only the latter is functional in math mode. But
by default text fonts are declared in mode node. Thus mathastext now
intervenes to make it so that the font it declares in math mode will
use mode basic. This fixes issues with for example old style figures
being used while the text font used lining figures (or vice versa,
depending on the font). But see the code comments for more.