mathastext 1.3y 2022-11-04
- re-organize documentation, mention the mathfont package and
  explain how to use frenchmath package with mathastext

- new options: ncccomma, binarysemicolon, frenchmath*

- under LGRgreek: \...up and \...it control sequences for all
  Greek letters, alphabets \mathgreekup and \mathgreekit.

  \MTgreekupdefault (with default 'n') replaces internal uses of
  \updefault which since LaTeX 2020-02-02 caused systematic 'up' to
  'n' substitutions with Font Warnings in the log.

The new LGRgreek features required an extensive internal refactoring
which is expected to not induce changes on most existing documents.  But
it may induce changes to those using some unusual advanced mathastext
configuration in the preamble.  For details and workarounds make sure to
read the dedicated section in the PDF documentation.