mathastext 1.4a 2024-07-20
With this release, your documents will compile faster: mathastext
does not use at all anymore the \everymath/\everydisplay registers
(except for a LuaLaTeX font matter) for the more advanced
functionalities it provides.

- option everymath will partially revert this core change; it is
  provided as a quick fix in case of problems arising in a
  document.  Under this option though, some bugfixes coming with
  this release are not applied.  This option will be removed
  at next major release.  Please report problems encountered
  without it before it gets removed.

- the internal infrastructure to use math active characters is
  made easier to tap into for the case of the ascii letters,
  allowing special tasks such as adding color or frames or
  use another font or whatever.

- option activedigits will let digits also be customizable
  as in previous item.

- the babel-spanish dot action in math mode is not lost anymore.

- a LuaLaTeX-only hack dating back to 2016 has been updated to
  handle fonts declared using Renderer=HarfBuzz.

- mathastext now requires LaTeX 2020-02-02 and has trimmed
  legacy branches supporting earlier releases; support for
  the EU1 and EU2 font encodings has been removed.

See the Change Log in the PDF for additional changes and details.
Some parts of the documentation have been extensively rewritten
or at least reordered, which always improves things.

There was a prior 1.4 the same day but it had a bug and 1.4a followed
immediately.