mathastext 1.4e 2024-10-26
Sadly the 1.4 release of 2024/07/20 had a regression which caused
input such as $f''$ to crash (but the `everymath' option which
re-enacts pre-1.4 legacy code could be used as a temporary
workaround).

It also broke the (not really recommended) option `unimathaccents'
for use with Unicode engines, and it caused \MTversion{normal}
to crash if with the `LGRgreek' option (and without `subdued').

This release fixes these problems and a few more obscure ones.
The package release script got updated to re-compile numerous very
ancient test files which would have, in particular, revealed the
$f''$ bug which occurred at 1.4.

Note: there was an 1.4d upload earlier that same day but two
bugs of earlier releases where discovered thereafter, which
warranted an 1.4e update.  See the bottom of mathastext.html
for details.